// src/types/aar.type.ts
export type TCar = {
   car_id: number;
   name: string;
   description: string;
   passengers: number;
   max_speed: string;
   gearbox_type: string;
   fuel_type: string;
   price_per_day: string;
   discount_percentage: number;
   image_url: string;
   created_at: string;
   updated_at: string;
};